fix bug

+++ coo.hxx
@@ -262,15 +262,15 @@ public:
         return true;
     }
 
-    int get_out_degree(size_t v){
-        int res = 0;
-        // size_t *num;
-        int *num;
+    size_t get_out_degree(size_t v){
+        size_t res = 0;
+        size_t *num;
+        // int *num;
         size_t vd;
         
         // memory allocation
-        // cudaMalloc((void **)&num, sizeof(size_t));
-        cudaMalloc((void **)&num, sizeof(int));
+        cudaMalloc((void **)&num, sizeof(size_t));
+        //cudaMalloc((void **)&num, sizeof(int));
         cudaMalloc((void **)&vd, sizeof(size_t));
         cudaMemcpy(num, res, sizeof(size_t), cudaMemcpyHostToDevice);
         cudaMemcpy(vd, v, sizeof(size_t), cudaMemcpyHostToDevice);
@@ -279,20 +279,20 @@ public:
         if (err != cudaSuccess) 
             printf("Error: %s\n", cudaGetErrorString(err));
         //bring data back 
-        // cudaMemcpy(res, num, sizeof(size_t), cudaMemcpyDeviceToHost);
-        cudaMemcpy(res, num, sizeof(int), cudaMemcpyDeviceToHost);
+        cudaMemcpy(res, num, sizeof(size_t), cudaMemcpyDeviceToHost);
+        // cudaMemcpy(res, num, sizeof(int), cudaMemcpyDeviceToHost);
         return res;
     }
 
-    int get_in_degree(size_t v){
-        int res = 0;
+    size_t get_in_degree(size_t v){
+        size_t res = 0;
         // size_t *num;
         int* num;
         size_t vd;
         
         // memory allocation
-        // cudaMalloc((void **)&num, sizeof(size_t));
-        cudaMalloc((void **)&num, sizeof(int));
+        cudaMalloc((void **)&num, sizeof(size_t));
+        // cudaMalloc((void **)&num, sizeof(int));
         cudaMalloc((void **)&vd, sizeof(size_t));
         cudaMemcpy(num, res, sizeof(size_t), cudaMemcpyHostToDevice);
         cudaMemcpy(vd, v, sizeof(size_t), cudaMemcpyHostToDevice);
@@ -301,8 +301,8 @@ public:
         if (err != cudaSuccess) 
             printf("Error: %s\n", cudaGetErrorString(err));
         //bring data back 
-        // cudaMemcpy(res, num, sizeof(size_t), cudaMemcpyDeviceToHost);
-        cudaMemcpy(res, num, sizeof(int), cudaMemcpyDeviceToHost);
+        cudaMemcpy(res, num, sizeof(size_t), cudaMemcpyDeviceToHost);
+        //cudaMemcpy(res, num, sizeof(int), cudaMemcpyDeviceToHost);
         return res;
     }
